server/requirements.txt
flask==3.0.0
plotly==5.18.0
pyserial==3.5